"added new functionality in the user controller and added test cases for that...

+++ src/main/java/com/depot/ims/repositories/UsersRepository.java
@@ -10,7 +10,7 @@ public interface UsersRepository extends JpaRepository<User, Long> {
     List<User> findAll();
 
     @Query("SELECT u FROM User u WHERE u.userId=?1")
-    User findByUserId(Integer userId);
+    User findByUserId(Long userId);
 
     @Query("SELECT u FROM User u WHERE u.username=?1")
     User findByUsername(String username);
